Implement LogReader.
The log of qemu process need to be consumed so that it won't be
suspended by os due to the continuous pipe write. The `LogReader`
is designed to solve this. It submits a task, which reads from the
provied file to the shared buffer continuously, to tokio runtime.
The shared buffer can be cleared, read to end via `LogReader`.